Adds Refactoring and divisiveness
