--- test/encrypt_test.rb
@@ -1,9 +1,0 @@
-require './test/test_helper'
-
-class EncryptTest < Minitest::Test
-
-  def setup
-    @enigma = Enigma.new
-  end
-
-end
