+++ test/test_helper.rb
@@ -6,4 +6,3 @@ require './lib/enigma'
 # require './lib/encryptor'
 require 'pry'
 require 'date'
-require 'matrix'
